i91: Move logic to useNewShout hook
