+++ react-app/src/library/components/BottomSheet/config.js
@@ -1,6 +1,6 @@
 import { normalizeValue } from 'res';
 
 export default Object.freeze({
-  ANIMATION_OFFSET: normalizeValue(80),
+  ANIMATION_OFFSET: normalizeValue(200),
   DISMISS_THRESHOLD: normalizeValue(80),
 });
